made some separate components where possible and made sure the functionality stayed the same
