+++ src/js/component/header.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Header = () => {
+  return (
+    <div>
+      <h1 className="display-2 text-secondary opacity-50">todos</h1>
+    </div>
+  );
+};
+
+export default Header;
